+++ lib/example/isi.dart
@@ -1,8 +1,0 @@
-import 'package:meta_model_transformer/meta/meta.dart';
-
-@MetaUse(field: 'data')
-class Isi<T>{
-  const Isi({this.data});
-
-  final dynamic data;
-}