+++ lib/generator/meta_model_transformer.dart
@@ -16,8 +16,11 @@ class SingleFileAggregateBuilder extends Builder {
   // with extension [.x.dart]
   String baseFileName = '';
 
-  // Base class meta count
-  int baseMetaCount = 0;
+  // Flag to check base class type
+  bool baseClassTypeIsAbstract = false;
+
+  // Counter meta
+  int checkMetaCount = 0;
 
   // Field meta value
   String fieldMetaValue = '';
@@ -53,9 +56,12 @@ class SingleFileAggregateBuilder extends Builder {
 ''');
 
     // Generated import
-    final String writePart = await _generateImport(assets, buildStep: buildStep);
+    final String writePart = await _generateImport(
+      assets,
+      buildStep: buildStep,
+    );
     buffer.write(writePart);
-    
+
     for (final AssetId file in assets) {
       // Guarding data
       // continue to next loop if data is not valid
@@ -92,29 +98,42 @@ class SingleFileAggregateBuilder extends Builder {
       }
     }
 
+    // Prevent generator when `MetaUse` annotation not found
+    ClassGuard.preventSomeConditionUsingMetaUseAnnotation(
+      requiredVariableName: fieldMetaValue,
+      baseMetaCount: checkMetaCount,
+    );
+
     // Generator
     buffer.write(await _writeGeneratorClass());
 
     // Write extension
-    buffer.write(writeExtension(
-      baseClass: baseFileName,
-      registry: registry,
-      fieldName: fieldMetaValue,
-    ));
-    
-    // Write to file
-    final output = AssetId(
-      buildStep.inputId.package,
-      'lib/generated/meta_transformer_generator.x.dart',
+    buffer.write(
+      writeExtension(
+        baseClass: baseFileName,
+        registry: registry,
+        fieldName: fieldMetaValue,
+      ),
     );
-    await buildStep.writeAsString(output, buffer.toString());
 
+    if (checkMetaCount == 0) {
+      return;
+    }
+
+    // Write to file
+    if (checkMetaCount != 0) {
+      final output = AssetId(
+        buildStep.inputId.package,
+        'lib/global/model/generated/meta_transformer_generator.x.dart',
+      );
+      await buildStep.writeAsString(output, buffer.toString());
+    }
     _hasRun = false;
   }
 
   @override
   Map<String, List<String>> get buildExtensions => {
-    r'$lib$': ['generated/meta_transformer_generator.x.dart'],
+    r'$lib$': ['global/model/generated/meta_transformer_generator.x.dart'],
   };
 
   /// Setup base meta model transform
@@ -134,30 +153,40 @@ class SingleFileAggregateBuilder extends Builder {
       buildStep: buildStep,
     );
 
-    if (!metaBaseIsEmpty) {
-      final AnnotatedElement firstAnnotation = dataMetaBase.first;
+    for (AnnotatedElement annotation in dataMetaBase) {
+      checkMetaCount++;
+
+      final ClassElement classElement = annotation.element as ClassElement;
 
       // Change name generator
-      baseFileName = firstAnnotation.element.name ?? '';
-
-      // Add counter
-      baseMetaCount++;
+      baseFileName = annotation.element.name ?? '';
 
       // Set field value
-      fieldMetaValue =
-          firstAnnotation.annotation.peek('field')?.stringValue ?? '';
+      fieldMetaValue = annotation.annotation.peek('field')?.stringValue ?? '';
+
+      // Check class type
+      if (classElement.isAbstract) {
+        baseClassTypeIsAbstract = true;
+        for (var e in classElement.constructors) {
+          if (e.formalParameters.isEmpty) {
+            throw InvalidGenerationSource(
+              'Constructor must have [$annotation as dynamic]',
+              todo:
+                  'Add field $fieldMetaValue into constuctor parameter as dynamic',
+              element: annotation.element,
+            );
+          }
+        }
+      }
 
       // Precent consumer has a field
-      if (!ClassGuard.isClassHasField(
-        fieldMetaValue,
-        data: firstAnnotation.element as ClassElement,
-      )) {
+      if (!ClassGuard.isClassHasField(fieldMetaValue, data: classElement)) {
         // Throw exception
         throw InvalidGenerationSource(
-          'Field $fieldMetaValue not found in ${firstAnnotation.element.name}',
+          'Field $fieldMetaValue not found in ${annotation.element.name}',
           todo:
-              'Add field $fieldMetaValue into ${firstAnnotation.element.name} as Map<String, dynamic>',
-          element: firstAnnotation.element,
+              'Add field $fieldMetaValue into ${annotation.element.name} as Map<String, dynamic>',
+          element: annotation.element,
         );
       }
     }
@@ -194,14 +223,14 @@ class SingleFileAggregateBuilder extends Builder {
       final String filePath = file.path.substring(4);
 
       if (!metaUseIsEmpty || !transformIsEmpty) {
-        partBuffer.writeln('import \'../$filePath\';');
+        partBuffer.writeln('import \'../../../$filePath\';');
       }
     }
 
     return partBuffer.toString();
   }
 
-  Future<StringBuffer> _writeGeneratorClass() async{
+  Future<StringBuffer> _writeGeneratorClass() async {
     final StringBuffer buffer = StringBuffer();
 
     buffer.writeln('''
@@ -209,7 +238,7 @@ class SingleFileAggregateBuilder extends Builder {
 typedef jsonParser = dynamic Function(dynamic);
 
 class _\$MetaModelTransformer {
-  static Map<String, jsonParser> registry = {
+  static Map<String, jsonParser> parseClosure = {
 
     // Registry for parsing List<T> objects.
     //