+++ lib/utils/class_guard.dart
@@ -1,9 +1,9 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:build/build.dart';
-import 'package:meta_model_transformer/meta/meta.dart';
 import 'package:source_gen/source_gen.dart';
 
 class ClassGuard {
+  ClassGuard._();
   static bool isClassHasField(String fieldName, {required ClassElement data}) {
     for (final FieldElement field in data.fields) {
       if (field.name == null || field.name == '') return false;
@@ -55,5 +55,18 @@ class ClassGuard {
     return true;
   }
 
-  final MetaUse data = MetaUse(field: 'data');
+  static void preventSomeConditionUsingMetaUseAnnotation({
+    required String requiredVariableName,
+    required int baseMetaCount,
+    Element? element,
+  }) {
+    switch (baseMetaCount) {
+      case > 1:
+        throw InvalidGenerationSource(
+          'You can not use @MetaUse more than once',
+          todo: 'Just select one @MetaUse annotation and remove the others',
+          element: element,
+        );
+    }
+  }
 }